moved into subpackage /imagestorage
