+++ backend/src/main/java/com/luiswiederhold/backend/flashcards/FlashcardService.java
@@ -2,13 +2,14 @@ package com.luiswiederhold.backend.flashcards;
 
 import com.luiswiederhold.backend.flashcards.DTOs.FlashcardSideDTO;
 import com.luiswiederhold.backend.flashcards.exception.LowConfidenceException;
+import com.luiswiederhold.backend.flashcards.imagestorage.ImageStorageService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.net.URI;
 import java.net.URISyntaxException;
-import java.nio.file.Path;
+
 // TODO: Username can't contain /
 @Service
 public class FlashcardService {
